@@ -10,51 +10,53 @@
     "generateZip": "set NODE_ENV=production && gulp generate-chrome-package"
   },
   "dependencies": {
-    "axios": "^0.15.3",
-    "office-ui-fabric-react": "^1.5.0",
-    "react": "^15.3.2",
-    "react-dom": "^15.3.2",
-    "react-router": "^3.0.2",
-    "redux-thunk": "2.1.1",
-    "sp-pnp-js": "^3.0.1"
+    "axios": "^1.10.0",
+    "office-ui-fabric-react": "^7.204.1",
+    "preact": "^10.26.9",
+    "react": "npm:@preact/compat",
+    "react-dom": "npm:@preact/compat",
+    "react-router": "^7.6.3",
+    "react-router-dom": "^7.6.3",
+    "redux-thunk": "3.1.0",
+    "sp-pnp-js": "^3.0.10"
   },
   "devDependencies": {
-    "@types/axios": "^0.9.35",
-    "@types/chrome": "0.0.35",
-    "@types/core-js": "^0.9.34",
-    "@types/microsoft-ajax": "0.0.31",
-    "@types/react": "^0.14.46",
-    "@types/react-dom": "^0.14.18",
-    "@types/react-redux": "^4.4.35",
-    "@types/react-router": "2.0.29",
-    "@types/redux-immutable-state-invariant": "^1.2.29",
-    "@types/redux-thunk": "2.1.32",
-    "@types/sharepoint": "^2013.1.2",
-    "awesome-typescript-loader": "^3.0.0-beta.18",
-    "css-loader": "^0.26.1",
-    "es6-promise": "4.0.5",
-    "extract-text-webpack-plugin": "^2.0.0-rc.2",
-    "gulp": "^3.9.1",
-    "gulp-replace": "^1.0.0",
+    "@types/axios": "^0.14.4",
+    "@types/bun": "^1.2.17",
+    "@types/chrome": "0.0.328",
+    "@types/core-js": "^2.5.8",
+    "@types/microsoft-ajax": "4.5.0",
+    "@types/react-redux": "^7.1.34",
+    "@types/redux-immutable-state-invariant": "^2.1.4",
+    "@types/redux-thunk": "2.1.0",
+    "@types/sharepoint": "^2016.1.14",
+    "awesome-typescript-loader": "^5.2.1",
+    "bun-dev-server": "^0.9.84",
+    "bun-dynamic-path": "^1.0.4",
+    "bun-style-plugin": "^1.4.0",
+    "css-loader": "^7.1.2",
+    "es6-promise": "4.2.8",
+    "extract-text-webpack-plugin": "^3.0.2",
+    "gulp": "^5.0.1",
+    "gulp-replace": "^1.1.4",
     "gulp-util": "^3.0.8",
-    "gulp-zip": "^3.2.0",
-    "http-server": "^0.9.0",
-    "immutable": "^3.8.1",
-    "node-sass": "^4.4.0",
-    "opener": "^1.4.2",
+    "gulp-zip": "^6.1.0",
+    "http-server": "^14.1.1",
+    "immutable": "^5.1.3",
+    "opener": "^1.5.2",
     "path": "^0.12.7",
-    "react-redux": "^5.0.1",
-    "redux": "^3.6.0",
-    "redux-immutable-state-invariant": "^1.2.4",
-    "sass-loader": "^4.1.1",
-    "source-map-loader": "^0.1.6",
-    "style-loader": "^0.13.1",
-    "tslint": "^4.4.2",
-    "tslint-react": "^2.3.0",
-    "typescript": "2.3.4",
-    "webpack": "^2.2.1",
-    "webpack-dev-server": "^2.2.1",
-    "whatwg-fetch": "1.0.0"
+    "react-redux": "^9.2.0",
+    "redux": "^5.0.1",
+    "redux-immutable-state-invariant": "^2.1.0",
+    "sass-loader": "^16.0.5",
+    "source-map-loader": "^5.0.0",
+    "style-loader": "^4.0.0",
+    "tslint": "^6.1.3",
+    "tslint-react": "^5.0.0",
+    "typescript": "5.8.3",
+    "webpack": "^5.99.9",
+    "webpack-dev-server": "^5.2.2",
+    "whatwg-fetch": "3.6.20"
   },
   "repository": {
     "type": "git",
